Use codeowner structure
Refactor the declarations menu to the correct codeowners structure
